+++ changelog.md
@@ -2,6 +2,7 @@
 ## v1.0.0-beta.4
 - <cg>Add</c> a very awful <cl>comment filtering system</c>
     - <c-a6a6a6>my first mistake was trusting dc with this mod</c>
+- <cg>Add</c> event level support (finally!)
 - <cg>Fix</c> custom ids
 
 ## v1.0.0-beta.3